Reorganize Launcher.Commands folder
