+++ src/Launcher/CLI/Contexts/MapCommandContext.cs
@@ -1,6 +1,6 @@
 ﻿using Launcher.Services;
 
-namespace Launcher.Commands;
+namespace Launcher.CLI.Contexts;
 
 internal abstract class MapCommandContext(string mapName)
 {
