+++ src/Launcher/CLI/Contexts/MapGenerateContext.cs
@@ -1,7 +1,7 @@
 ﻿using AutoMapper;
 using Launcher.Services;
 
-namespace Launcher.Commands;
+namespace Launcher.CLI.Contexts;
 
 internal sealed class MapGenerateContext : MapCommandContext
 {
