+++ src/Launcher/CLI/Contexts/MapSerializationContext.cs
@@ -1,7 +1,7 @@
 ﻿using AutoMapper;
 using Launcher.Services;
 
-namespace Launcher.Commands;
+namespace Launcher.CLI.Contexts;
 
 internal sealed class MapSerializationContext(string mapName) : MapCommandContext(mapName)
 {
