--- src/Launcher/Commands/MapOutputKind.cs
@@ -1,7 +1,0 @@
-﻿namespace Launcher.Commands;
-
-internal enum MapOutputKind
-{
-  Directory,
-  File
-}
